#565 Added get() function to the state_trees module
